Stage 2c cover fetch
Fetching the cover was actually pretty easy. Puzzling out a layout to display it nicely was considerably more difficult.

Added BookResultView to show the title, author and cover but also handle partial information. Updated the scan screen to use the book view. Added preview support to scan and search models, so we could have comprehensive previews for the book and scan screen. This was super helpful, fixed a bug where scan progress wasn't displayed.