Stage 3 prep: book screen, app icon and launch screen
Added a book screen. The scan screen "add" button now saves the book in memory. Renamed and update the home screen to show the book screen when there are saved books.

Some ui polish, to make buttons and lists use body font. We'll settle on body mostly throughout - we'll just let the user customize the app with their dynamic text setting, large or small.

We will commit the book screen for fun, but we won't be keeping it as is. Really we want the home screen to be a searchable author list. Our persistence schema just got a bit more complicated.