Stage 3: author and book persistence
A bigger change than expected, partly due to learning swift data. Partly due to the realization that author will be rather more prominent and must be its own entity. The landing screen really does need to be a list of authors.

The key new object is StorageModel, which configures swift data and implements the storage of books and authors. Also implements a preview model that supports swift ui previews, though this will need further polish.

The rest of the app was touched by this shake up: renames, moves and adjustments.