Stage 3: added edition and cover art cache
Scanning two editions of Stormbringer highlighted why open library distinguishes between editions and works. Editions have an ISBN and cover art and can be owned, but "have I read" and ratings really apply to ALL editions, the work. Feels important to get this right and make it look simple.

So a rather large change to split an Edition entity out of the rough Book entity. Many adjustments to the models and their previews, and a big change to the book screen so it shows all editions of the book.

These two refactors broke the image on the result view. Introduced the cover model as an environment object to cache cover images. This turned out really well, it was easy to move the image properties to views from the edition, which is really a better spot. The edition now just has cover ids.

All this is just begging for a nice adjustment to the scan flow but best to stop making this change larger. The unit tests have really been lagging too.