Stage 4: appearance and flow polish
Sweeping style changes to improve appearance and adaptive layout. Refactored view modifiers for consistency and moved them to a new style group. Refactored several new smaller views so they could be composed easily.

The heart of this change is the new scanned book screen. The layout looks much, much better with the cover art large and on top, and the buttons grouped at the bottom. And with the new composable smaller views and a few uses of viewThatFits, the layout also works much better in landscape and even on ipad.

Finally a few changes to the scan screen and view. The uikit data scanner just does not play nicely with a navigation view: the camera stalls on return. There is probably a way to get this working since a background/foreground can restart the camera. Tried programmatic navigation but it's just so much more work than showing a sheet. So next up, show the scanner and scanned book screen as sheets. The camera won't lock up because we'll never be returning to the scan screen.