Stage 4, flow polish: scanner fix and book refactor
The main change here is showing the scanner and "approve scanned book" screens as sheets. I'd wanted to keep the scanner in the navigation view to save us a click during a multiple scan session. But the camera stops when we pop back to the scanner. Interestingly it restarts if we background/foreground the app but that isn't a great solution. It's probably best to move to single use scanner screens anyway, those were bound to crash eventually.

The part of the change that is bigger than it looks is a great refactor: adding a book struct. Renamed the old book entity to "work" which matches open library. Added a new book struct which wraps together all the bits that make up a real physical book: edition, work and authors. You can see from files touched just how many places in the code needed and were greatly simplified by the new book struct.