Stage 4, flow polish: updated browse book screen
Updated the work (formerly book) screen to match the awesome new layout of the scanned book screen. The new edition page view worked out really well. It's just a simple tab view. Very easy to drop in several editions, scroll, and communicate selected edition back to the parent work screen. All with the view that fits layout that adapts nicely on landscape. Another touch to get the horizontal layout on ipad and this will be just about perfect.

Added another author and a book with two editions to test the multi edition list in previews.