Added delete button and editable modifier
Added a delete button. It uses the platform destructive appearance but adds font, border and shape customizations to match the custom "bookmind" button style. Added a button to the work screen but did not yet reconnect it with a delete handler.

Particularly pleased with the new editable modifier. This fully adapts the complicated edit mode environment variable to the simple boolean flag I expect. Very easy to drop onto any view, and has an optional change handler as well.