Fixes #6 added edition delete handling
Added edition delete to the work screen. If it is the last edition of a work, we delete the work as well, and pop back to the author screen. Seems presumptive to delete the author, so we'll leave that for the user to decide.

Refactored work layout back and forth a bit, and landed on an any layout that should work nicely for phone. And should be easy to adjust for ipad and mac. Much simpler than the previous duplicated views, or even a separate group view. Updated both the work and scanned book screens with the any layout.