Added search button and rough screen
Added a search button to the home screen, above scan. The search button opens the rough search screen. The search screen is a rough layout, the search button is not yet functional. Used a nice library background, might use that on other screens too.

Added a "found text" state to the scan model. The coordinator sets this state when the scanner has found text, but not an isbn. We use this state to show a different help message. Because in practice, we're usually in this state because something is wrong with the isbn.