Fixes #18 add found text state
In the previous commit we'd started a search screen where the user can type an ISBN. It's a good change and going well but out of scope for this ticket. The search screen is going to need some larger refactors, and the stated purpose of this ticket and branch is achieved. When the search screen is in we'll see if we can add a button to navigate to it directly from the "having trouble scanning" error state.