29 added search router with mixed results
Moved the home screens flags and insert book properties to a new type, SearchRouter. Nice simple little observable class that can be passed in and used by search and scan views as just one property. But then the home screen can observe those properties and push the right screens. This works, even to pop and replace, as with scan/search, scan/insert and search/insert.

But! No animation. When we pop that extra screen the replacement just instantly blasts onscreen, no nice sliding. So we'll park this in case we want to come back to this approach. But we WILL park for now and try to get navigation path to work again.

It's annoying to have to define an entire type just to push the search, scan and insert screens, but maybe if we keep those types very small...