29 fixed search flow animations
In surprising results, flipping two modal toggles does *not* animate correctly, but popping two navigation views does animate correctly. So much for showing the search screen as a modal, for consistency. That doesn't animate when we switch to search from a troublesome scan.

So the search navigation is a bit of a confused mess but feels somewhat right. With this change we've broken inserting the selected book to core data, but we have other plans for that anyway so will deal with it in the next commit.