29 add search by isbn screen
Added InsertBookModel, to restore saving selected scan/search results. We start with just the plain de-dupe and save we had, but the model gives us a place to add cancel support. Updated the insert screen (soon to be edit screen?) and search progress view to use the new model.

In testing, noticed it would be nice if we could delete an author from the author screen. Also nice if the author was deleted when we deleted their last work. So added those.

This fixes #29.