7 update search to first check local storage
Fixes #7. Before searching openlibrary, we do a query for stored editions with the same isbn. Interesting... note to self, check searching for isbn 10 vs. isbn 13. If we have a stored edition, scan/search screens return that instead of searching.

Another note, think of a nice place to put that shared snippet. Maybe a view modifier? Tricky to add to the search model because only views have access to the storage environment variable.

Deleted the insert book model, replacing it with a couple static methods on Book and Edition. Updated navigation to use the Book type instead of the InsertBookModel type. We had to pass so much in to the model that it didn't really encapsulate much.

Finally updated the insert book screen. Surprisingly easy for scan/search screens to just run a quick query for the isbn. Removed the rating view from the insert screen: it's going away and we were changing that area. Updated the insert screen to show save and cancel buttons only if the edition isn't saved. Removed the delete button, that will only be available while browring.