24 added several scanning fixes
Fixes #24,  fixes #11. Added support and tests for a few more variations of ISBN with spaces and colons. Discovered a fix for some older 9 digit sbns which were not found: just convert to 10 digits.

Added another fix for a weird bug with using saved results. Since editions have always been saved with the isbn 13 from lookup, a 10 digit id manually entered in the search screen is treated as a new book. Fixed this by always converted to 13 digits before a local lookup.