Fixes #35 add isbn example to search field prompt text
The search screen text field now includes an example isbn in its prompt text. Hopefully it demonstrates that you don't *need* to type the dashes, although you still can.

Added an "Invalid" state to the search model. The search screen sets this state when the user types a number that is not an isbn. The search screen will show a message about valid isbn format.

Updated the search model message for "not found". The message now includes a link to open library and lets the user know they can add the book there. The message also includes a teaser about a manual entry screen.