Fixes #36 delete author after deleting their last book
This is already implemented on the author screen. But now we also have the book screen delete the author. We rely on the author screen to check it has no books and pop itself... this may be problematic later.

Snuck in a surprisingly large refactor where we moved the search router to an environment variable. Big change but simplified a ton of code.