Fixes #40 add book title edit text field
A title text field now appears when the book screen edit button is tapped. The field is bound to the work title property, so changing the text updates the book title. Snuck in a list of authors that also appears when edit is tapped, read only for now. We'll add behavior to the list in the "fix author" ticket.

Added constructors to the work and author screens so that the author and work properties can be private, but also preview can show the editing layouts.

Removed subtitle property from work. The data I've seen has been inconsistent and low utility. Added an ISBN 13 label to the work screen.