14 add author filters: converted pickers to menus
We'll be adding a picker to the home screen so the user can pick a filter. The picker is inconsistent since it doesn't allow bold text like the two existing buttons. Tried alerts, action sheets and menus, which all allow customization of the opening "button", but *none* of the resulting dialogs respect dynamic text size. Decided on menu as the dialog is at least more concise without extraneous cancel button or title text.

So in preparation for the new filter menu button, converted the existing read and own state buttons to menus.