14 add library filters
Added library filter. It is a struct with a title, so it may be shown in a menu in the home screen. But also it provides a function to filter a list of authors, based on the own or read state of the authors works and editions. So now you can filter to see the books you own or have read or want to read.

Added a filter menu view to the project, and added an instance to the home screen, as a toolbar button. Updated the library screen to use the filter.